refactor: VideoFile과 Video의 역할 분리
